+++ trunk/rg/sketch_line.c
@@ -77,7 +77,7 @@ edit(void *p)
 }
 
 static void
-mousebuttondown(void *p, RG_Sketch *sk, double x, double y, int button)
+mousebuttondown(void *p, RG_Sketch *sk, float x, float y, int button)
 {
 	struct line_tool *lt = p;
 	VG *vg = sk->vg;
@@ -119,8 +119,7 @@ mousebuttondown(void *p, RG_Sketch *sk, double x, double y, int button)
 }
 
 static void
-mousemotion(void *p, RG_Sketch *sk, double x, double y, double xrel,
-    double yrel)
+mousemotion(void *p, RG_Sketch *sk, float x, float y, float xrel, float yrel)
 {
 	struct line_tool *lt = p;
